Restructured Validation code packaging a bit to simplify the design pattern, with a new TRAPIResponseValidator wrapper.
