+++ tests/test_validation_report.py
@@ -7,6 +7,22 @@ from reasoner_validator.report import CodeDictionary, ValidationReporter
 TEST_TRAPI_VERSION = "1.3.0"
 TEST_BIOLINK_VERSION = "2.4.8"
 
+def check_messages(validator: ValidationReporter, code):
+    messages: Dict[str, List[Dict]] = validator.get_messages()
+    if code:
+        # TODO: 'code' should be found in code.yaml
+        # value: Optional[Tuple[str, str]] = CodeDictionary._code_value(code)
+        # assert value is not None
+        message_type = validator.get_message_type(code)
+        if message_type == "error":
+            assert any([error['code'] == code for error in messages['errors']])
+        elif message_type == "warning":
+            assert any([warning['code'] == code for warning in messages['warnings']])
+        elif message_type == "info":
+            assert any([info['code'] == code for info in messages['information']])
+    else:  # no errors or warnings expected? Assert absence of such messages?
+        assert not validator.has_messages(), f"Unexpected messages seen {messages}"
+
 
 def test_message_loader():
     assert CodeDictionary._code_value("") is None
